% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeProp.R
\name{computeProp}
\alias{computeProp}
\title{Perform reference-free deconvolution to estimate cell-type proportion}
\usage{
computeProp(DataType = "Gene expression", Y.raw, refinx, K)
}
\arguments{
\item{DataType}{Character specifying data type. Must be either:
\itemize{
\item "Gene expression" (default): Uses deconfounding via \code{deconf} package
\item "DNA methylation": Uses Reference-Free EWAS via \code{RefFreeEWAS} package
}}

\item{Y.raw}{A raw data matrix of complex samples where rows represent features (genes/CpG sites)
and columns represent samples.}

\item{refinx}{estimated cell-type specific features}

\item{K}{the number of cell type}
}
\value{
estimated cell-type proportion matrix (K x samples), where rows
correspond to cell types and columns to samples.
}
\description{
Perform reference-free deconvolution to estimate cell-type proportion
}
