% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEKtissue.R
\name{DEKTissue}
\alias{DEKTissue}
\title{Cross-cell type differential analysis}
\usage{
DEKTissue(
  K,
  Y,
  Prop,
  design,
  WhichPar = NULL,
  contrast.vec = NULL,
  sort = F,
  var.threshold = 0.1,
  data.threshold = TRUE
)
}
\arguments{
\item{K}{the number of cell types}

\item{Y}{a raw data matrix of complex samples}

\item{Prop}{cell-type proportions for all subjects}

\item{design}{a design matrix representing the status of samples (or continuous features: need to be developed)}

\item{WhichPar}{a number chosen from 1 to 2K (which parameter will be tested?)}

\item{contrast.vec}{contrast.vec should be a 2K*1 vector to specify a contrast vector}

\item{sort}{whether the p value is sorted}

\item{var.threshold}{var.threshold controls how much you want to bound your variation estimation. Default value is 0.1}

\item{data.threshold}{whether to bound varBeta}
}
\value{
A dataframe containing p-value, q-value and statistics
}
\description{
Cross-cell type differential analysis
}
