% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVarSelect_2VSother.R
\name{DEVarSelect_2VSother}
\alias{DEVarSelect_2VSother}
\title{Cross-cell type differential analysis between two cell types and the other cell types (named as 2_O)}
\usage{
DEVarSelect_2VSother(Y.raw, Prop0, nMarker = 1000)
}
\arguments{
\item{Y.raw}{A raw data matrix of complex samples}

\item{Prop0}{estimated cell type proportion matrix}

\item{nMarker}{the number of cell type specific markers}
}
\value{
estimated cell-type specific markers
}
\description{
Cross-cell type differential analysis between two cell types and the other cell types (named as 2_O)
}
