% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVarSelect_pairwise.R
\name{DEVarSelect_pairwise}
\alias{DEVarSelect_pairwise}
\title{Cross-cell type differential analysis between one cell types and another (named as P_W)}
\usage{
DEVarSelect_pairwise(Y.raw, Prop0, nMarker = 1000)
}
\arguments{
\item{Y.raw}{A raw data matrix of complex samples}

\item{Prop0}{estimated cell type proportion matrix}

\item{nMarker}{number of cell type specific markers}
}
\value{
estimated cell type specific markers
}
\description{
Cross-cell type differential analysis between one cell types and another (named as P_W)
}
