% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVarSelect_RFdecd.R
\name{DEVarSelect_RFdecd}
\alias{DEVarSelect_RFdecd}
\title{Cross-cell type differential analysis between one cell type and the other cell types
as well as between two cell types and the other cell types (named RFdecd)}
\usage{
DEVarSelect_RFdecd(Y.raw, Prop0, nMarker = 1000)
}
\arguments{
\item{Y.raw}{A raw data matrix of complex samples}

\item{Prop0}{estimated cell type proportion matrix}

\item{nMarker}{number of cell type specific markers}
}
\value{
estimated cell type specific markers
}
\description{
Cross-cell type differential analysis between one cell type and the other cell types
as well as between two cell types and the other cell types (named RFdecd)
}
