% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVarSelect_VAR.R
\name{DEVarSelect_VAR}
\alias{DEVarSelect_VAR}
\title{select the top 1000 features with the largest variance in the estimated cell-type profiles (named VAR)}
\usage{
DEVarSelect_VAR(Y.raw, Prop0, nMarker = 1000)
}
\arguments{
\item{Y.raw}{A raw data matrix of complex samples}

\item{Prop0}{estimated cell type proportion matrix}

\item{nMarker}{number of cell type specific markers}
}
\value{
estimated cell type specific markers
}
\description{
select the top 1000 features with the largest variance in the estimated cell-type profiles (named VAR)
}
