% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRefinxVAR.R
\name{findRefinx.VAR}
\alias{findRefinx.VAR}
\title{find index for cell-type specific features based on the largest variation in the raw data}
\usage{
findRefinx.VAR(rawdata, nMarker = 1000)
}
\arguments{
\item{rawdata}{A raw data matrix from complex samples.}

\item{nMarker}{Number of cell-type specific features.}
}
\value{
The index of top features with the largest variation.
}
\description{
find index for cell-type specific features based on the largest variation in the raw data
}
