% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeProp.R
\name{computeProp}
\alias{computeProp}
\title{Perform reference-free deconvolution to estimate cell-type proportion}
\usage{
computeProp(DataType = "Gene expression", Y.raw, refinx, K)
}
\arguments{
\item{DataType}{either 'Gene expression' or 'DNA methylation'}

\item{Y.raw}{A raw data matrix of complex samples}

\item{refinx}{estimated cell-type specific features}

\item{K}{the number of cell type}
}
\value{
estimated cell-type proportion matrix
}
\description{
Perform reference-free deconvolution to estimate cell-type proportion
}
