% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVarSelect_1VSother.R
\name{DEVarSelect_1VSother}
\alias{DEVarSelect_1VSother}
\title{Cross-cell type differential analysis between one cell type and the other cell types (named as 1_O)}
\usage{
DEVarSelect_1VSother(Y.raw, Prop0, nMarker = 1000)
}
\arguments{
\item{Y.raw}{a raw data matrix of complex samples}

\item{Prop0}{estimated cell-type proportion matrix}

\item{nMarker}{the number of cell type specific markers}
}
\value{
estimated cell-type specific markers
}
\description{
Cross-cell type differential analysis between one cell type and the other cell types (named as 1_O)
}
