% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVarSelect_CV.R
\name{DEVarSelect_CV}
\alias{DEVarSelect_CV}
\title{Select the top 1000 features with the largest coefficient of variation in the estimated cell-type profiles (named CV)}
\usage{
DEVarSelect_CV(Y.raw, Prop0, nMarker = 1000)
}
\arguments{
\item{Y.raw}{A raw data matrix of complex samples}

\item{Prop0}{estimated cell type proportion matrix}

\item{nMarker}{number of cell type specific markers}
}
\value{
estimated cell type specific markers
}
\description{
Select the top 1000 features with the largest coefficient of variation in the estimated cell-type profiles (named CV)
}
