% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFdeconv.R
\name{RFdeconv}
\alias{RFdeconv}
\title{Feature-selection-based reference-free deconvolution method}
\usage{
RFdeconv(
  DataType = "Gene expression",
  Y_raw,
  K,
  CTSoption = DEVarSelect_1VSother,
  nMarker = 1000,
  InitMarker = NULL,
  TotalIter = 30
)
}
\arguments{
\item{DataType}{either 'Gene expression' or 'DNA methylation'}

\item{Y_raw}{A raw data matrix of complex samples}

\item{K}{Number of cell type}

\item{CTSoption}{feature selection options}

\item{nMarker}{number of cell type specific markers}

\item{InitMarker}{initial cell type specific markers}

\item{TotalIter}{number of iterations}
}
\value{
A list containing rmse,proportions with 30 iterations,
the estimated proportion matrix corresponding to the iteration with the smallest RMSE
}
\description{
Feature-selection-based reference-free deconvolution method
}
